Não desenhando quando o size e zero, colocando completion

+++ GenerativeArt/Template/PlaygroundBook/Sources/Canvas/Pixel.swift
@@ -10,7 +10,7 @@ import Foundation
 public struct Pixel {
     
     public typealias Position = Int
-    public typealias Color = Int
+    public typealias Color = Float
     
     public var position: Position
     public var color: Color
